refactor: updated the blogs post
